Fix classpath in manifest when auto-build is disabled
m2e resolves workspace projects to target/classes, or pom.xml when the build directory doesn't exist.
When generating the Classpath entry in the MANIFEST.MF, the current way of dealing with workspace
dependencies is to fake a jar when the project's artifact.file is mapped to a directory.

When auto-build is disabled, target/classes is not necessarily generated hence the manifest contains
 pom.xml entries in lieu of artifactId.jar's

The fix simply extends the fake jar mechanism to replace pom.xml entries when computing the Manifest classpath.

Signed-off-by: Fred Bricon <fbricon@gmail.com>